apps/paciente/serializers.py
from rest_framework import serializers

from apps.paciente.models import Paciente


class PacienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Paciente
        fields = '__all__'
