requirements.txt
Django==3.1.7

django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djoser==2.1.0
social-auth-app-django==4.0.0
django-rest-auth==0.9.5
django-allauth==0.52.0
django-ckeditor==6.1.0
dj-rest-auth==3.0.0
pytz==2021.3

Pillow==9.2.0
mutagen==1.45.1 
psycopg2-binary==2.9.5

argon2-cffi==21.1.0 
django-environ==0.4.5

stripe==2.58.0
braintree==4.13.1

#pyJWT==2.1.0
PyJWT==2.6.0

boto3==1.18.63
django-storages==1.12.2 

markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.3.0

requests==2.26.0

django-anymail==8.6
mailjet-rest==1.3.4
django-cloudinary-storage==0.3.0
google-auth==2.16.2
algoliasearch-django==2.0.0

#djangorestframework-simplejwt==5.2.2

python-decouple==3.8
